+++ src/Reflection/AbstractMethodEntity.php
@@ -111,15 +111,15 @@ abstract class AbstractMethodEntity extends AbstractEntity
      * Returns `@throws` tags
      * @return array
      */
-    public function getThrowsTags(): array
-    {
-        return array_map(function (TagEntity $tag) {
-            return [
-                'type' => ltrim((string)$tag->getType(), '\\'),
-                'description' => $tag->getDescription(),
-            ];
-        }, $this->getTagsByName('throws'));
-    }
+//    public function getThrowsTags(): array
+//    {
+//        return array_map(function (TagEntity $tag) {
+//            return [
+//                'type' => ltrim((string)$tag->getType(), '\\'),
+//                'description' => $tag->getDescription(),
+//            ];
+//        }, $this->getTagsByName('throws'));
+//    }
 
     /**
      * Returns `true` if it's static