+++ CHANGELOG.md
@@ -7,8 +7,9 @@
 * added `ClassEntity::createFromName()` static method;
 * added `TestCase::assertStringEqualsTemplate()` assertion method;
 * added template `class` element;
-* removed `DeprecatedTrait::getDeprecatedDescription()` and
-    `DeprecatedTrait::isDeprecated()` method. Use instead the `TagEntity`;
+* removed `DeprecatedTrait::getDeprecatedDescription()`,
+    `DeprecatedTrait::isDeprecated()` and `AbstractMethodEntity::getThrowsTags()`
+    methods. Use instead the `TagEntity` and its methods;
 * updated `bootstrap` and `highlight` asset files.
 
 ### 1.0.1-beta2